+++ varhab-cli/src/main/java/com/github/bihealth/varhab/VarhabException.java
@@ -1,8 +1,6 @@
 package com.github.bihealth.varhab;
 
-/**
- * Thrown in case of problems.
- */
+/** Thrown in case of problems. */
 public class VarhabException extends Exception {
   public VarhabException(String message) {
     super(message);