+++ CHANGES.md
@@ -5,6 +5,6 @@
 - [ ] implement `annotate`
     - [ ] proper small variant genotypes output
     - [ ] proper variant-specific annotation
-- [ ] implement `init-db`
-    - [ ] ExAC TSV
-    - [ ] ClinVar TSV (MacArthur style)
+- [x] implement `init-db`
+    - [x] ExAC TSV
+    - [x] ClinVar TSV (MacArthur style)